refactor(data-warehouse): move person-property consumer into logic
The product structure check only recognizes canonical backend directories
(facade, presentation, logic, models, ...); backend/consumers/ was not one, so
it bypassed import-linter enforcement and failed Repo checks. Move the consumer
and its tests into backend/logic/ (alongside the existing person-property sync
logic, matching how other products place Kafka consumers) and update the
management command import.

Generated-By: PostHog Code
Task-Id: 652d5991-61f5-4d1b-92cc-a29757e9932d
