--- products/customer_analytics/backend/consumers/__init__.py
@@ -1,1 +1,0 @@
-"""Consumers owned by customer_analytics."""
